requirements.txt
flask>=3.0.0
waitress>=3.0.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
supabase>=2.0.0
requests>=2.31.0
anthropic>=0.40.0
google-genai>=1.0.0
scikit-learn>=1.4.0
numpy>=1.24.0

# --- local analysis pipeline (not used on Render) ---
yt-dlp>=2024.12.13
faster-whisper>=1.0.0
scenedetect[opencv]>=0.6.4
librosa>=0.10.0
Pillow>=10.0.0
